test: refactor so that all vm related tests can be reused
This commit reshuffles the code a bit so that we a reusable
./test/vmtest directory that can be used by the images library.

With that we can add a toplevel pyproject.toml file so that
we can import vmtest via
```console
$ pip install git+https@github.com/osbuild/bootc-image-builder
```
im other projects.

Note that none of this is ideal, butt this is (hopefully)
a temporary measure until we find a more permanent home for
our vm runner or replace it with something like test.thing or
the new osbuild QEMU code.
