+++ test/conftest.py
@@ -1,6 +1,8 @@
 import pytest
 
+# pylint: disable=wrong-import-order
 from testcases import TestCase
+from vmtest.util import get_free_port
 
 
 def pytest_addoption(parser):
@@ -20,3 +22,8 @@ def pytest_make_parametrize_id(config, val):  # pylint: disable=W0613
     if isinstance(val, TestCase):
         return f"{val}"
     return None
+
+
+@pytest.fixture(name="free_port")
+def free_port_fixture():
+    return get_free_port()
