+++ vmtest/util_test.py
@@ -1,10 +1,11 @@
 import contextlib
-import platform
+import shutil
 import subprocess
 from unittest.mock import call, patch
 
 import pytest
-from testutil import get_free_port, has_executable, wait_ssh_ready
+
+from vmtest.util import get_free_port, wait_ssh_ready
 
 
 def test_get_free_port():
@@ -12,20 +13,17 @@ def test_get_free_port():
     assert 1024 < port_nr < 65535
 
 
-@pytest.fixture(name="free_port")
-def free_port_fixture():
-    return get_free_port()
-
-
 @patch("time.sleep")
-def test_wait_ssh_ready_sleeps_no_connection(mocked_sleep, free_port):
+def test_wait_ssh_ready_sleeps_no_connection(mocked_sleep):
+    free_port = get_free_port()
     with pytest.raises(ConnectionRefusedError):
         wait_ssh_ready("localhost", free_port, sleep=0.1, max_wait_sec=0.35)
     assert mocked_sleep.call_args_list == [call(0.1), call(0.1), call(0.1)]
 
 
-@pytest.mark.skipif(not has_executable("nc"), reason="needs nc")
-def test_wait_ssh_ready_sleeps_wrong_reply(free_port):
+@pytest.mark.skipif(not shutil.which("nc"), reason="needs nc")
+def test_wait_ssh_ready_sleeps_wrong_reply():
+    free_port = get_free_port()
     with contextlib.ExitStack() as cm:
         with subprocess.Popen(
             f"echo not-ssh | nc -vv -l -p {free_port}",
@@ -47,12 +45,3 @@ def test_wait_ssh_ready_sleeps_wrong_reply(free_port):
                     wait_ssh_ready("localhost", free_port, sleep=0.1, max_wait_sec=0.55)
                 assert mocked_sleep.call_args_list == [
                     call(0.1), call(0.1), call(0.1), call(0.1), call(0.1)]
-
-
-@pytest.mark.skipif(platform.system() == "Darwin", reason="hangs on macOS")
-@pytest.mark.skipif(not has_executable("nc"), reason="needs nc")
-def test_wait_ssh_ready_integration(free_port):
-    with contextlib.ExitStack() as cm:
-        with subprocess.Popen(f"echo OpenSSH | nc -l -p {free_port}", shell=True) as p:
-            cm.callback(p.kill)
-            wait_ssh_ready("localhost", free_port, sleep=0.1, max_wait_sec=10)
